--- a/aten/src/ATen/mkldnn/Runtime.cpp
+++ b/aten/src/ATen/mkldnn/Runtime.cpp
@@ -1,5 +1,0 @@
-#include <ATen/mkldnn/Runtime.h>
-
-namespace at { namespace native {
-
-}}  // namespace at::native