--- a/aten/src/ATen/native/mkldnn/IDeepRegistration.cpp
+++ b/aten/src/ATen/native/mkldnn/IDeepRegistration.cpp
@@ -1,8 +1,21 @@
+#include <ATen/ATen.h>
 #include <ATen/Config.h>
 
 #if AT_MKLDNN_ENABLED()
 
 // needs to be included only once in library.
 #include <ideep_pin_singletons.hpp>
+
+using namespace ideep;
+
+RegisterEngineAllocator cpu_alloc(
+  engine::cpu_engine(),
+  [](size_t size) {
+    return c10::GetAllocator(c10::DeviceType::CPU)->raw_allocate(size);
+  },
+  [](void* p) {
+    c10::GetAllocator(c10::DeviceType::CPU)->raw_deallocate(p);
+  }
+);
 
 #endif // AT_MKLDNN_ENALBED()