--- a/aten/src/ATen/native/mkldnn/Relu.cpp
+++ b/aten/src/ATen/native/mkldnn/Relu.cpp
@@ -26,14 +26,14 @@ namespace at { namespace native {
 Tensor mkldnn_relu(const Tensor& input) {
   const ideep::tensor& x = itensor_from_mkldnn(input);
   ideep::tensor y;
-  ideep::eltwise_forward::compute<AllocForMKLDNN>(
+  ideep::eltwise_forward::compute(
       x, y, ideep::algorithm::eltwise_relu, ideep::prop_kind::forward_training, /*alpha*/ 0.0);
   return new_with_itensor_mkldnn(std::move(y), input.options());
 }
 
 Tensor& mkldnn_relu_(Tensor& input) {
   ideep::tensor& x = itensor_from_mkldnn(input);
-  ideep::eltwise_forward::compute<AllocForMKLDNN>(
+  ideep::eltwise_forward::compute(
       x, x, ideep::algorithm::eltwise_relu, ideep::prop_kind::forward_training, /*alpha*/ 0.0);
   return input;
 }