--- a/aten/src/ATen/native/mkldnn/TensorShape.cpp
+++ b/aten/src/ATen/native/mkldnn/TensorShape.cpp
@@ -50,7 +50,7 @@ Tensor mkldnn_reshape(const Tensor& self, IntArrayRef size) {
   }
   const ideep::tensor& x = itensor_from_mkldnn(self);
   ideep::tensor y{x};
-  y.reshape<AllocForMKLDNN>({inferred_size.cbegin(), inferred_size.cend()});
+  y.reshape(inferred_size);
   return new_with_itensor_mkldnn(std::move(y), self.options());
 }
 
@@ -61,7 +61,7 @@ Tensor mkldnn_clone(const Tensor& self, c10::optional<c10::MemoryFormat> optiona
       optional_memory_format.value());
   ideep::tensor& src = itensor_from_mkldnn(self);
   ideep::tensor dst;
-  ideep::direct_copy::compute<AllocForMKLDNN>(src, dst);
+  ideep::direct_copy::compute(src, dst);
   return new_with_itensor_mkldnn(std::move(dst), self.options());
 }
 
@@ -71,7 +71,7 @@ Tensor mkldnn_transpose(const Tensor & self, int64_t dim0, int64_t dim1) {
   std::vector<int> axes(x.ndims());
   std::iota(axes.begin(), axes.end(), 0);
   std::swap(axes[dim0], axes[dim1]);
-  y.transpose_from<AllocForMKLDNN>(x, axes);
+  y.transpose_from(x, axes);
   return new_with_itensor_mkldnn(std::move(y), self.options());
 }