--- a/caffe2/ideep/operators/adam_op.cc
+++ b/caffe2/ideep/operators/adam_op.cc
@@ -110,11 +110,11 @@ class IDEEPAdamOp final : public IDEEPOperator {
     CAFFE_ENFORCE(grad.get_nelems() == moment_1.get_nelems());
     CAFFE_ENFORCE(grad.get_nelems() == moment_2.get_nelems());
     if (params != *out_params)
-        out_params->reinit(params.get_descriptor());
+        out_params->init(params.get_descriptor());
     if (moment_1 != *out_moment1)
-        out_moment1->reinit(moment_1.get_descriptor());
+        out_moment1->init(moment_1.get_descriptor());
     if (moment_2 != *out_moment2)
-        out_moment2->reinit(moment_2.get_descriptor());
+        out_moment2->init(moment_2.get_descriptor());
     const auto w = static_cast<float *>(params.get_data_handle());
     const auto g = static_cast<float *>(grad.get_data_handle());
     const auto m = static_cast<float *>(moment_1.get_data_handle());
@@ -146,7 +146,7 @@ class IDEEPAdamOp final : public IDEEPOperator {
     } else {
       auto* out_grad = Output(OUTPUT_GRAD);
       if (grad != *out_grad)
-        out_grad->reinit(grad.get_descriptor());
+        out_grad->init(grad.get_descriptor());
       auto ng = static_cast<float *>(out_grad->get_data_handle());
       adam_ideep_compute_output_grad(
           grad.get_nelems(),