@vadimkantorov 2020-01-20
Should then PyTorch use jemalloc by default?