@CaoZhongZ 2020-01-21
The performance degradation caused by allocator is large allocation eagerly come from and return to kernel. Page clean imposed by kernel has the complexity of a ReLU, linear to tensor size.

We'll have a more detailed tech gist to explain more.