@Jianhui-Li 2020-01-22
@gottbrath @VitalyFedyunin This is the DNNL upgrade. Our future BFloat16 work depends on this PR. Appreciate if you can review it and give some feedback.