@jgong5 2020-01-23
> Should then PyTorch use jemalloc by default?

We had some related discussions from a separate PR FYI: https://github.com/pytorch/pytorch/pull/20567#issuecomment-511621826
So far, preloading jemalloc is the best thing we can do. @dzhulgakov had the concern to make it by default due to packaging concerns: https://github.com/pytorch/pytorch/pull/20567#issuecomment-513599884