@gchanan 2020-01-27
can you fix the merge conflicts?

Also, there is an MKL-DNN rebrand to DNNL but the updated environment variables (e.g. MKLDNN_BUILD_EXAMPLES) still use the MKLDNN name?