@vpirogov 2020-01-27
@gchanan,

Old environment variable names and build options continue to work with DNNL v1.2 to maintain backward compatibility. These will be removed in the next major version though.