@gchanan 2020-01-27
@vpirogov: I'm asking about the _new_ environment variables (see the Old/New table), which have the MKLDNN name.