@dzhulgakov 2020-01-27
Regarding jemalloc - if we want to use it for entire python process, we'd need to either compile `python` with it or preload the library with LD_PRELOAD. If we care only about tensor memory though, it should be possible to just invoke jemalloc APIs directly and wrap that into c10's allocator. The downside is that we'd have two malloc implementations competing for resources.

Which model is most affected by it? Seems like ShuffleNet? Is there a one line repro?