@pinzhenx 2020-01-28
Thank you for taking the time to review this PR.

<br>

> @gchanan: I'm asking about the new environment variables (see the Old/New table), which have the MKLDNN name.

You are correct, these env variables and build options now have new names with the prefix `DNNL_`. But as @vpirogov mentioned, old names still continue to work for backward compatibility.

As for the rebranding, we first need to substitute all MKLDNN/MKL-DNN/mkldnn/mkl-dnn with DNNL/dnnl in the repo, and also need to take care of public APIs like `to_mkldnn()` and other stuff like tensor layout `mkldnn`. We hope this PR not to be overwhelmed by tons of renaming and new APIs, so we kept most of the old names to make this upgrade as smooth as possible, except for few of them in `FindMKLDNN.cmake` where renaming has to be made.

<br>

> @dzhulgakov: Which model is most affected by it? Seems like ShuffleNet? Is there a one line repro?
> If we care only about tensor memory though, it should be possible to just invoke jemalloc APIs directly and wrap that into c10's allocator. The downside is that we'd have two malloc implementations competing for resources.

From existing models we have, we only observe the caffe2 shufflenet has been affected in **throughput test**. I'm sorry we cannot provide you with the repro for now, I might have one for you someday later.

Also, we don't think jemalloc is a must for now, I highlighted it only to point out one caveat of this upgrade, that in some heavy workloads throughput performance may degrade, while the scope of the effects is very limited.