@Coderx7 2020-02-02
@pinzhenx, @Jianhui-Li Thanks a lot for this, This was really needed. great job. 
By the way, when are we going to have OpenCL integrated as well? 
Does it require significant changes to the codebase in order to use DNNL's GPU acceleration?