@pinzhenx 2020-02-25
Hi @peterjc123, we are encountering a build error on the Windows platform, however, we are unable to repro it on our side with vs2019. May I please ask for your kind assistance with this issue?

And hi @vpirogov, it seems that this issue occurs in the DNNL CMake files. If you have some ideas, please also let me know, thanks!

Thank you in advance for your help.