@pinzhenx 2020-03-18
Here's my results from Skylake
```
GCC: gcc version 7.5.0 (Ubuntu 7.5.0-3ubuntu1~18.04)
OS: Ubuntu 18.04.4 LTS
CPU: Intel(R) Xeon(R) Platinum 8180 CPU @ 2.50GHz
RAM: 187 G
Config: 50warm 500run 1T
```

ms per iter | new 1.5.0a0+94a4015 | old 1.5.0a0+67608cc
-- | -- | --
resnext50 | 83.85 | 98.32
resnext101 | 149.66 | 159.65
resnext152 | 214.27 | 220.55



<br>

... and this one is from Cascade Lake
```
GCC: gcc version 7.3.1 20180303 (Red Hat 7.3.1-5) (GCC)
OS: CentOS Linux 7 (Core)
CPU: Intel(R) Xeon(R) Platinum 8280L CPU @ 2.70GHz
RAM: 187G
Config: 50warm 500run 1T
```


ms per iter | new 1.5.0a0+94a4015 | old 1.5.0a0+67608cc
-- | -- | --
resnext50 | 78.16 | 93.76
resnext101 | 137.04 | 151.24
resnext152 | 197.08 | 211.38


<br>

However, I was not able to reproduce your performance numbers. From the pytorch profiling results, I observed that the difference was mainly saved in batchnorm, and the conv time were almost the same. 

Could you please provide some other info about your system so that I could find a more similar machine to reproduce it later?

-------

I was comparing this PR with the one you cherrypicked to fix group conv and prepack weights.
```
new:
94a4015e1d6ab8e895dfac606a2ed062d22ed134
fix reshape issue of blocked tensor
 
old:
67608cc018b4548fcccfad2d9969c7d546224736
Fix MKLDNN conv2d 5d weight handling (#34115)
```